bot_package/requirements.txt
discord.py==2.5.2
aiohttp==3.9.2
requests==2.31.0
python-dotenv==1.0.0
